Re #1 Basic UI framework using ReactJs
A Single Page Application framework using ReactJs. Broadly, its organized into two
containers -

1. a "home" container (home page) which shows summary of stuff which is not tenant specific,
   a list of gateways and list of tenants are part of the home container today

2. a "tenant" container which lists stuff specific to a tenant - like users and application
   bundles and policies etc..

The expectation is that the user will first go to the home page and see the summary and then
click on a particular tenant and get into the second tenant container view

The containers are all in src/containers. The fine grained features inside the container is
all in src/views .. The src/views/home has all the features in the home container and the
src/views/tenant/ has all the per-tenant features like viewing/updating users and app bundles
etc..