Re #6 Add a UI/UX page to input certificate information
Right now we used self signed certs obviously cause we have not purchased
from any signing authority etc.. So agent needs to know what the CA cert
is when establishing ssl with the gateway. Add a page where we can enter
multiple certificates so that it gets sent to the controller and controller
provides it to agent as part of onboarding

Also added a temporary mechanism to pin an agent to a gateway - useful for
testing becuase we want to predictably assign agents and connectors to
gateways while testing. So the user/bundle page has an extra field gateway
which can be left empty in which case the gateway will be picked as per
the best gateway selection algorithm